refactor(component) : Navbar and Input
add new general component input and change header to navbar
