+++ src/App.js
@@ -1,14 +1,14 @@
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
 import "./App.scss";
-import Header from "./components/header";
+import Navbar from "./components/navbar";
 import Home from "./pages/home";
 
 function App() {
   return (
     <Router>
       <div className="app">
-        <Header />
+        <Navbar />
         <div className="app__page-container">
           <div className="app__page">
             <Switch>
