+++ src/components/input/index.scss
@@ -1,0 +1,14 @@
+.input {
+  padding: 0.8rem 1.6rem;
+  box-shadow: 0px 10px 20px rgba(0, 0, 0, 0.04), 0px 2px 6px rgba(0, 0, 0, 0.04),
+    0px 0px 1px rgba(0, 0, 0, 0.04);
+  border-radius: 10px;
+  border: 2px solid #dfdede;
+  transition: all 0.2s ease;
+  &:focus {
+    outline: none;
+    border: 2px solid var(--primary);
+    box-shadow: 0px 10px 20px rgba(75, 198, 160, 0.04),
+      0px 2px 6px rgba(75, 198, 160, 0.04), 0px 0px 1px rgba(75, 198, 160, 0.04);
+  }
+}
