+++ src/components/navbar/index.scss
@@ -7,6 +7,7 @@
   position: fixed;
   top: 0;
   width: 100%;
+  z-index: 1000;
   &__title {
     font-size: 2.8rem;
     color: var(--primary);
