+++ src/index.scss
@@ -1,6 +1,6 @@
 :root {
-  --primary: #4BC6A0;
-  --primaryFont: 'Asap', sans-serif;
+  --primary: #4bc6a0;
+  --primaryFont: "Asap", sans-serif;
 }
 
 * {
